add credential-types credential tab
